requirements.txt
numpy==1.20.1
Pillow==8.1.2
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.7.1+cu110
-f https://download.pytorch.org/whl/torch_stable.html 
torchaudio==0.7.2
-f https://download.pytorch.org/whl/torch_stable.html 
torchvision==0.8.2+cu110
typing-extensions==3.7.4.3
